+++ src/RcppExports.cpp
@@ -26,6 +26,69 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// lambda_matrix_loop
+arma::mat lambda_matrix_loop(arma::mat xq, arma::mat xe, arma::vec xbq, arma::vec xbe, arma::vec G1_prime_xq, arma::vec G1_prime_prime_xq, arma::vec G2_xe, arma::vec G2_prime_xe, arma::vec G2_prime_prime_xe, arma::vec density, arma::vec cdf, double alpha, bool include_misspecification_terms);
+RcppExport SEXP _esreg_lambda_matrix_loop(SEXP xqSEXP, SEXP xeSEXP, SEXP xbqSEXP, SEXP xbeSEXP, SEXP G1_prime_xqSEXP, SEXP G1_prime_prime_xqSEXP, SEXP G2_xeSEXP, SEXP G2_prime_xeSEXP, SEXP G2_prime_prime_xeSEXP, SEXP densitySEXP, SEXP cdfSEXP, SEXP alphaSEXP, SEXP include_misspecification_termsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type xq(xqSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type xe(xeSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type xbq(xbqSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type xbe(xbeSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type G1_prime_xq(G1_prime_xqSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type G1_prime_prime_xq(G1_prime_prime_xqSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type G2_xe(G2_xeSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type G2_prime_xe(G2_prime_xeSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type G2_prime_prime_xe(G2_prime_prime_xeSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type density(densitySEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type cdf(cdfSEXP);
+    Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
+    Rcpp::traits::input_parameter< bool >::type include_misspecification_terms(include_misspecification_termsSEXP);
+    rcpp_result_gen = Rcpp::wrap(lambda_matrix_loop(xq, xe, xbq, xbe, G1_prime_xq, G1_prime_prime_xq, G2_xe, G2_prime_xe, G2_prime_prime_xe, density, cdf, alpha, include_misspecification_terms));
+    return rcpp_result_gen;
+END_RCPP
+}
+// sigma_matrix_loop
+arma::mat sigma_matrix_loop(arma::mat xq, arma::mat xe, arma::colvec xbq, arma::colvec xbe, arma::colvec G1_prime_xq, arma::colvec G2_xe, arma::colvec G2_prime_xe, arma::colvec conditional_variance, arma::vec cdf, double alpha, bool include_misspecification_terms);
+RcppExport SEXP _esreg_sigma_matrix_loop(SEXP xqSEXP, SEXP xeSEXP, SEXP xbqSEXP, SEXP xbeSEXP, SEXP G1_prime_xqSEXP, SEXP G2_xeSEXP, SEXP G2_prime_xeSEXP, SEXP conditional_varianceSEXP, SEXP cdfSEXP, SEXP alphaSEXP, SEXP include_misspecification_termsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type xq(xqSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type xe(xeSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type xbq(xbqSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type xbe(xbeSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type G1_prime_xq(G1_prime_xqSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type G2_xe(G2_xeSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type G2_prime_xe(G2_prime_xeSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type conditional_variance(conditional_varianceSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type cdf(cdfSEXP);
+    Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
+    Rcpp::traits::input_parameter< bool >::type include_misspecification_terms(include_misspecification_termsSEXP);
+    rcpp_result_gen = Rcpp::wrap(sigma_matrix_loop(xq, xe, xbq, xbe, G1_prime_xq, G2_xe, G2_prime_xe, conditional_variance, cdf, alpha, include_misspecification_terms));
+    return rcpp_result_gen;
+END_RCPP
+}
+// estimating_function_loop
+arma::mat estimating_function_loop(arma::vec y, arma::mat xq, arma::mat xe, arma::colvec xbq, arma::colvec xbe, arma::colvec G1_prime_xq, arma::colvec G2_xe, arma::colvec G2_prime_xe, double alpha);
+RcppExport SEXP _esreg_estimating_function_loop(SEXP ySEXP, SEXP xqSEXP, SEXP xeSEXP, SEXP xbqSEXP, SEXP xbeSEXP, SEXP G1_prime_xqSEXP, SEXP G2_xeSEXP, SEXP G2_prime_xeSEXP, SEXP alphaSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::vec >::type y(ySEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type xq(xqSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type xe(xeSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type xbq(xbqSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type xbe(xbeSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type G1_prime_xq(G1_prime_xqSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type G2_xe(G2_xeSEXP);
+    Rcpp::traits::input_parameter< arma::colvec >::type G2_prime_xe(G2_prime_xeSEXP);
+    Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
+    rcpp_result_gen = Rcpp::wrap(estimating_function_loop(y, xq, xe, xbq, xbe, G1_prime_xq, G2_xe, G2_prime_xe, alpha));
+    return rcpp_result_gen;
+END_RCPP
+}
 // G1_fun
 double G1_fun(double z, int type);
 RcppExport SEXP _esreg_G1_fun(SEXP zSEXP, SEXP typeSEXP) {
@@ -47,6 +110,18 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type z(zSEXP);
     Rcpp::traits::input_parameter< int >::type type(typeSEXP);
     rcpp_result_gen = Rcpp::wrap(G1_prime_fun(z, type));
+    return rcpp_result_gen;
+END_RCPP
+}
+// G1_prime_prime_fun
+double G1_prime_prime_fun(double z, int type);
+RcppExport SEXP _esreg_G1_prime_prime_fun(SEXP zSEXP, SEXP typeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< double >::type z(zSEXP);
+    Rcpp::traits::input_parameter< int >::type type(typeSEXP);
+    rcpp_result_gen = Rcpp::wrap(G1_prime_prime_fun(z, type));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -86,6 +161,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// G2_prime_prime
+double G2_prime_prime(double z, int type);
+RcppExport SEXP _esreg_G2_prime_prime(SEXP zSEXP, SEXP typeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< double >::type z(zSEXP);
+    Rcpp::traits::input_parameter< int >::type type(typeSEXP);
+    rcpp_result_gen = Rcpp::wrap(G2_prime_prime(z, type));
+    return rcpp_result_gen;
+END_RCPP
+}
 // G_vec
 Rcpp::NumericVector G_vec(Rcpp::NumericVector z, Rcpp::String g, int type);
 RcppExport SEXP _esreg_G_vec(SEXP zSEXP, SEXP gSEXP, SEXP typeSEXP) {
@@ -119,11 +206,16 @@ END_RCPP
 
 static const R_CallMethodDef CallEntries[] = {
     {"_esreg_l_esreg_covariance", (DL_FUNC) &_esreg_l_esreg_covariance, 10},
+    {"_esreg_lambda_matrix_loop", (DL_FUNC) &_esreg_lambda_matrix_loop, 13},
+    {"_esreg_sigma_matrix_loop", (DL_FUNC) &_esreg_sigma_matrix_loop, 11},
+    {"_esreg_estimating_function_loop", (DL_FUNC) &_esreg_estimating_function_loop, 9},
     {"_esreg_G1_fun", (DL_FUNC) &_esreg_G1_fun, 2},
     {"_esreg_G1_prime_fun", (DL_FUNC) &_esreg_G1_prime_fun, 2},
+    {"_esreg_G1_prime_prime_fun", (DL_FUNC) &_esreg_G1_prime_prime_fun, 2},
     {"_esreg_G2_curly_fun", (DL_FUNC) &_esreg_G2_curly_fun, 2},
     {"_esreg_G2_fun", (DL_FUNC) &_esreg_G2_fun, 2},
     {"_esreg_G2_prime_fun", (DL_FUNC) &_esreg_G2_prime_fun, 2},
+    {"_esreg_G2_prime_prime", (DL_FUNC) &_esreg_G2_prime_prime, 2},
     {"_esreg_G_vec", (DL_FUNC) &_esreg_G_vec, 3},
     {"_esreg_esr_rho_lp", (DL_FUNC) &_esreg_esr_rho_lp, 7},
     {NULL, NULL, 0}