+++ src/target_functions.cpp
@@ -49,6 +49,30 @@ double G1_prime_fun(double z, int type) {
 
 
 //' @title Specification Function
+//' @description G1_prime_prime
+//' @param z Data
+//' @param type Choice of the G1_prime_prime function:
+//' \itemize{
+//'   \item 1: G1_prime(z) = 0
+//'   \item 2: G1_prime(z) = 0
+//' }
+//' @keywords internal
+//' @export
+// [[Rcpp::export]]
+double G1_prime_prime_fun(double z, int type) {
+  double out;
+  if (type == 1) {
+    out = 0;
+  } else if (type == 2) {
+    out = 0;
+  } else {
+    Rcpp::stop("type not in 1, 2!");
+  }
+  return out;
+}
+
+
+//' @title Specification Function
 //' @description G2_curly
 //' @param z Data
 //' @param type Choice of the G2_curly function:
@@ -164,6 +188,46 @@ double G2_prime_fun(double z, int type) {
   return out;
 }
 
+
+//' @title Specification Function
+//' @description G2_prime_prime
+//' @param z Data
+//' @param type Choice of the G2_prime_prime function:
+//' \itemize{
+//'   \item 1: -2/z^3, z < 0
+//'   \item 2: 0.375 / (-z)^(5/2), z < 0
+//'   \item 3: 6/z^4, z < 0
+//'   \item 4: -(exp(z) * (exp(z) - 1)) / (exp(z) + 1)^3
+//'   \item 5: exp(z)
+//' }
+//' @keywords internal
+//' @export
+// [[Rcpp::export]]
+double G2_prime_prime(double z, int type) {
+  double out;
+  if ((type == 1) | (type == 2) | (type == 3)) {
+    if (z >= 0) {
+      Rcpp::warning("z can not be positive for type 1, 2, 3!");
+      return NA_REAL;
+    }
+  }
+  if (type == 1) {
+    out = -1/pow(z, 3);
+  } else if (type == 2) {
+    out = 0.375/pow(-z, 2.5);
+  } else if (type == 3) {
+    out = 6/pow(z, 4);
+  } else if (type == 4) {
+    out =  -(exp(z) * (exp(z) - 1)) / pow(1 + exp(z), 3);
+  } else if (type == 5) {
+    out = exp(z);
+  } else {
+    Rcpp::stop("type not in 1, 2, 3, 4, 5!");
+  }
+  return out;
+}
+
+
 //' @title Vectorized call to the G1 / G2 functions
 //' @description Vectorized call to the G1 / G2 functions
 //' @param z Vector
@@ -180,12 +244,16 @@ Rcpp::NumericVector G_vec(Rcpp::NumericVector z, Rcpp::String g, int type) {
     for (int i = 0; i < n; i++) out[i] = G1_fun(z[i], type);
   } else if (g == "G1_prime") {
     for (int i = 0; i < n; i++) out[i] = G1_prime_fun(z[i], type);
+  } else if (g == "G1_prime_prime") {
+    for (int i = 0; i < n; i++) out[i] = G1_prime_prime_fun(z[i], type);
   } else if (g == "G2_curly") {
     for (int i = 0; i < n; i++) out[i] = G2_curly_fun(z[i], type);
   } else if (g == "G2") {
     for (int i = 0; i < n; i++) out[i] = G2_fun(z[i], type);
   } else if (g == "G2_prime") {
     for (int i = 0; i < n; i++) out[i] = G2_prime_fun(z[i], type);
+  } else if (g == "G2_prime_prime") {
+    for (int i = 0; i < n; i++) out[i] = G2_prime_prime(z[i], type);
   } else {
     Rcpp::stop("Non supported G-function!");
   }